requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
DateTime==5.5
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
idna==3.10
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipdeptree==2.26.1
platformdirs==4.3.8
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
VirtualEnvironment==0.1.2
zope.interface==7.2
